+++ AvgScript/src/extension.ts
@@ -4,7 +4,7 @@ import { labelDefinition, labelReference } from './functions/label';
 import { hover, hoverFile } from './functions/hover';
 import { inlayHint } from './functions/inlayHint';
 import { diagnosticsCollection, onUpdate, triggerUpdate } from './functions/diagnostic';
-import { assetsListPanel, commandBasePath, commandBasePath_impl, commandGetAssetsList, commandGetAssetsList_impl, commandRefreshAssets, commandRefreshAssets_impl, commandShowJumpFlow, commandShowJumpFlow_impl, commandUpdateCommandExtension, commandUpdateCommandExtension_impl } from './functions/command';
+import { assetsListPanel, commandBasePath, commandBasePath_impl, commandGetAssetsList, commandGetAssetsList_impl, commandRefreshAssets, commandRefreshAssets_impl, commandReplaceScript, commandReplaceScript_impl, commandShowJumpFlow, commandShowJumpFlow_impl, commandUpdateCommandExtension, commandUpdateCommandExtension_impl } from './functions/command';
 import { outline } from './functions/outline';
 import { rename } from './functions/rename';
 import { atCommands, fileName, fileSuffix, langPrefix, settingsParam, sharpCommands } from './functions/completion';
@@ -64,6 +64,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	vscode.commands.registerCommand(commandUpdateCommandExtension, commandUpdateCommandExtension_impl);
 	vscode.commands.registerCommand(commandGetAssetsList, commandGetAssetsList_impl);
 	vscode.commands.registerCommand(commandShowJumpFlow, commandShowJumpFlow_impl);
+	vscode.commands.registerCommand(commandReplaceScript, commandReplaceScript_impl);
 
 	//--------------------
 	// Goto Definition